refactor(gfql/polars): pattern-existence family -> pattern_apply.py (module-per-family, Leo steer)
row_pipeline.py is the expression/projection lowering core; feature families get
their own modules (degrees.py precedent). Pure move + import updates.
